+++ flumine/backtest/backtest.py
@@ -1,5 +1,6 @@
 import logging
 import datetime
+from collections import defaultdict
 
 from ..baseflumine import BaseFlumine
 from ..events import events
@@ -32,21 +33,72 @@ class FlumineBacktest(BaseFlumine):
             )
         with self:
             self._monkey_patch_datetime()
-
+            """
+            list of either single stream or complete events depending 
+            on event_processing flag:
+               single: {None: [<Stream 1>], [<Stream 2>, ..]}
+               event: {123: [<Stream 1>, <Stream 2>, ..], 456: [..]}
+            Event data to be muxed/processed chronologically as per 
+            live rather than single which is per market in isolation.
+            """
+            event_streams = defaultdict(list)  # eventId: [<Stream>, ..]
             for stream in self.streams:
-                stream_gen = stream.create_generator()
+                event_id = stream.event_id if stream.event_processing else None
+                event_streams[event_id].append(stream)
 
-                logger.info(
-                    "Starting historical market '{0}'".format(stream.market_filter)
-                )
-                for event in stream_gen():
-                    self._process_market_books(events.MarketBookEvent(event))
-
-                self.handler_queue.clear()
-
-                logger.info(
-                    "Completed historical market '{0}'".format(stream.market_filter)
-                )
+            for event_id, streams in event_streams.items():
+                if event_id and len(streams) > 1:
+                    logger.info(
+                        "Starting historical event '{0}'".format(event_id),
+                        extra={
+                            "event_id": event_id,
+                            "markets": [s.market_filter for s in streams],
+                        },
+                    )
+                    # create cycles
+                    cycles = []  # [[epoch, [MarketBook], gen], ..]
+                    for stream in streams:
+                        stream_gen = stream.create_generator()()
+                        market_book = next(stream_gen)
+                        publish_time_epoch = market_book[0].publish_time_epoch
+                        cycles.append([publish_time_epoch, market_book, stream_gen])
+                    # process cycles
+                    while cycles:
+                        # order by epoch
+                        cycles.sort(key=lambda x: x[0])
+                        # get current
+                        _, market_book, stream_gen = cycles.pop(0)
+                        # process current
+                        self._process_market_books(events.MarketBookEvent(market_book))
+                        # gen next
+                        try:
+                            market_book = next(stream_gen)
+                        except StopIteration:
+                            continue
+                        publish_time_epoch = market_book[0].publish_time_epoch
+                        # add back
+                        cycles.append([publish_time_epoch, market_book, stream_gen])
+                    self.handler_queue.clear()
+                    logger.info("Completed historical event '{0}'".format(event_id))
+                else:
+                    for stream in streams:
+                        logger.info(
+                            "Starting historical market '{0}'".format(
+                                stream.market_filter
+                            ),
+                            extra={
+                                "market": stream.market_filter,
+                            },
+                        )
+                        stream_gen = stream.create_generator()
+                        for event in stream_gen():
+                            self._process_market_books(events.MarketBookEvent(event))
+                        self.handler_queue.clear()
+                        logger.info(
+                            "Completed historical market '{0}'".format(
+                                stream.market_filter
+                            )
+                        )
 
             self._process_end_flumine()
 
@@ -60,9 +112,9 @@ class FlumineBacktest(BaseFlumine):
             market_id = market_book.market_id
             config.current_time = market_book.publish_time
 
-            # check if there are orders to process
+            # check if there are orders to process (limited to current market only)
             if self.handler_queue:
-                self._check_pending_packages()
+                self._check_pending_packages(market_id)
 
             if market_book.status == "CLOSED":
                 self._process_close_market(event=events.CloseMarketEvent(market_book))
@@ -102,24 +154,30 @@ class FlumineBacktest(BaseFlumine):
         orders through strategies
         """
         blotter = market.blotter
+        completed_orders = []
         for order in blotter.live_orders:
             process.process_current_order(order)
             if order.trade.status == TradeStatus.COMPLETE:
-                blotter.complete_order(order)
+                completed_orders.append(order)
+        for order in completed_orders:
+            blotter.complete_order(order)
         for strategy in self.strategies:
             strategy_orders = blotter.strategy_orders(strategy)
             strategy.process_orders(market, strategy_orders)
 
-    def _check_pending_packages(self):
+    def _check_pending_packages(self, market_id: str) -> None:
         processed = []
         for order_package in self.handler_queue:
-            if order_package.elapsed_seconds > order_package.simulated_delay:
+            if (
+                order_package.market_id == market_id
+                and order_package.elapsed_seconds > order_package.simulated_delay
+            ):
                 order_package.client.execution.handler(order_package)
                 processed.append(order_package)
         for p in processed:
             self.handler_queue.remove(p)
 
-    def _monkey_patch_datetime(self):
+    def _monkey_patch_datetime(self) -> None:
         config.current_time = datetime.datetime.utcnow()
 
         class NewDateTime(datetime.datetime):
@@ -130,7 +188,7 @@ class FlumineBacktest(BaseFlumine):
         self._old_datetime = datetime.datetime
         datetime.datetime = NewDateTime
 
-    def _unpatch_datetime(self):
+    def _unpatch_datetime(self) -> None:
         datetime.datetime = self._old_datetime
 
     def __repr__(self) -> str: